cleaned code and fixed formatting
