+++ gopher/display_data.py
@@ -1,14 +1,11 @@
 from .annotations import load_annotations
 import pandas as pd
 import logging
-from tqdm.auto import tqdm
 from .stats import rankdata
 import numpy as np
 import matplotlib.pyplot as plt
 from scipy import stats
 import seaborn as sns
-
-LOGGER = logging.getLogger(__name__)
 
 
 def map_proteins(
