+++ tests/unit_tests/normalize_test.py
@@ -1,17 +1,5 @@
-import pytest
-from pathlib import Path
 import pandas as pd
 from gopher import normalize
-
-
-@pytest.fixture
-def real_data(tmp_path):
-    """Test using small files."""
-    fasta_df = Path("../data/small-yeast.fasta")
-    quant = pd.read_csv("../data/yeast_small.csv")
-    quant = quant.set_index("Protein")
-
-    return quant, fasta_df
 
 
 def test_normalize(real_data):
